feat(errors): добавить библиотеку shared-errors и передачу userId через grpc metadata
- создать пакет @clarte/shared-errors с чистыми доменными, grpc и http ошибками
- реализовать мапперы кодов и сущностей ошибок между слоями
- обновить proto-схемы и контроллеры сервисов на передачу userId через gRPC metadata
- добавить модуль Node в api-gateway

+++ apps/api-gateway/src/app/todo/infrastructure/clients/todo.client.ts
@@ -3,6 +3,7 @@ import { OnModuleInit } from '@nestjs/common';
 import { type ClientGrpc } from '@nestjs/microservices';
 import { Todo } from '@clarte/shared-contracts/proto';
 import { map, Observable } from 'rxjs';
+import { makeGrpcMetadata } from '@clarte/shared-nest/functions';
 
 export class TodoClient implements ITodoClient, OnModuleInit {
   private todoService!: Todo.TodoServiceClient;
@@ -13,27 +14,31 @@ export class TodoClient implements ITodoClient, OnModuleInit {
     this.todoService = this.todoGrpcClient.getService(Todo.TODO_SERVICE_NAME);
   }
 
-  createTodo(data: Todo.CreateTodoRequest): Observable<Todo.CreateTodoResponse> {
-    return this.todoService.createTodo(data);
+  createTodo(userId: string, data: Todo.CreateTodoRequest): Observable<Todo.CreateTodoResponse> {
+    return this.todoService.createTodo(data, makeGrpcMetadata({ userId }));
   }
 
-  updateTodo(data: Todo.UpdateTodoRequest): Observable<void> {
-    return this.todoService.updateTodo(data).pipe(map(() => void 0));
+  updateTodo(userId: string, data: Todo.UpdateTodoRequest): Observable<void> {
+    return this.todoService.updateTodo(data, makeGrpcMetadata({ userId })).pipe(map(() => void 0));
   }
 
-  completeTodo(data: Todo.CompleteTodoRequest): Observable<void> {
-    return this.todoService.completeTodo(data).pipe(map(() => void 0));
+  completeTodo(userId: string, data: Todo.CompleteTodoRequest): Observable<void> {
+    return this.todoService
+      .completeTodo(data, makeGrpcMetadata({ userId }))
+      .pipe(map(() => void 0));
   }
 
-  uncompleteTodo(data: Todo.UncompleteTodoRequest): Observable<void> {
-    return this.todoService.uncompleteTodo(data).pipe(map(() => void 0));
+  uncompleteTodo(userId: string, data: Todo.UncompleteTodoRequest): Observable<void> {
+    return this.todoService
+      .uncompleteTodo(data, makeGrpcMetadata({ userId }))
+      .pipe(map(() => void 0));
   }
 
-  deleteTodo(data: Todo.DeleteTodoRequest): Observable<void> {
-    return this.todoService.deleteTodo(data).pipe(map(() => void 0));
+  deleteTodo(userId: string, data: Todo.DeleteTodoRequest): Observable<void> {
+    return this.todoService.deleteTodo(data, makeGrpcMetadata({ userId })).pipe(map(() => void 0));
   }
 
   getUserTodos(userId: string): Observable<Todo.GetUserTodsResponse> {
-    return this.todoService.getUserTodos({ id: '', userId });
+    return this.todoService.getUserTodos({}, makeGrpcMetadata({ userId }));
   }
 }
